python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visa-direct-sdk"
version = "0.3.0"
description = "Visa Direct SDK (Python)"
authors = [{ name = "Visa Direct Team" }]
requires-python = ">=3.10"
dependencies = [
	"requests>=2.31.0",
	"jwcrypto>=1.5.6",
	"opentelemetry-api>=1.23.0",
	"opentelemetry-sdk>=1.23.0",
	"boto3>=1.34.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["visa_direct_sdk*"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "coverage>=7.2.0"]
